requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-mysql==3.11.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.2
psycopg2==2.8.6
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
yarl==1.6.3
